Fix `WorkflowOutStream` behaviour
This commit fixes a regression in the `WorkflowOutStream` behaviour.
Since the `cell_id` is stored in a `ContextVar`, it is necessary to
explicitly propagate it when scheduling work for the `pub_thread`.

This commit introduces a wrapper class that automatically passes the
`cell_id` value to the other thread when calling the `_flush` method.